Moved utility functions to Utils.kt
