+++ frontend/src/main/kotlin/HomeScreen.kt
@@ -1,11 +1,6 @@
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ExitToApp
@@ -15,7 +10,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
@@ -130,32 +124,6 @@ fun HomeScreen(userId: String, onNavigate: (Screen) -> Unit, onSetParty: (Party)
 }
 
 @Composable
-fun YesNoDialog(question: String, yesAnswer: String, noAnswer: String, onAccept: () -> Unit, onDismiss: () -> Unit) {
-    Dialog(onDismissRequest = onDismiss) {
-        Box(
-            modifier = Modifier
-                .wrapContentSize(Alignment.Center)
-        ) {
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier
-                    .background(color = MaterialTheme.colors.background, RoundedCornerShape(16.dp))
-                    .padding(16.dp)
-            ) {
-                Text(text = question)
-                Spacer(modifier = Modifier.height(16.dp))
-                Button(onClick = onAccept) {
-                    Text(yesAnswer)
-                }
-                Button(onClick = onDismiss) {
-                    Text(noAnswer)
-                }
-            }
-        }
-    }
-}
-
-@Composable
 fun LogoutDialog(onDismiss: () -> Unit, onLogout: () -> Unit) {
     YesNoDialog(
         "Do you really want to log out?",
@@ -180,82 +148,6 @@ fun RemoveFriendDialog(onDismiss: () -> Unit, userId: String, friend: Account) {
         },
         onDismiss
     )
-}
-
-@Composable
-fun TextFieldDialog(question: String, onAccept: (String) -> Unit, onDismiss: () -> Unit) {
-    Dialog(onDismissRequest = onDismiss) {
-        var text by remember { mutableStateOf("") }
-        Box(
-            modifier = Modifier
-                .wrapContentSize(Alignment.Center)
-        ) {
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier
-                    .background(color = MaterialTheme.colors.background, RoundedCornerShape(16.dp))
-                    .padding(16.dp)
-            ) {
-                Text(text = question)
-                TextField(
-                    value = text,
-                    onValueChange = { text = it },
-                    label = { Text("Name") }
-                )
-                Button(onClick = { onAccept(text) }) {
-                    Text("Accept")
-                }
-                Button(onClick = onDismiss) {
-                    Text("Cancel")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ListHeader(name: String, showButton: Boolean, buttonDescription: String = "", onButtonClick: () -> Unit = {}) {
-    Row(
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier.padding(16.dp).fillMaxWidth()
-    ) {
-        Text(
-            text = name,
-            fontSize = 24.sp
-        )
-        if (showButton) {
-            IconButton(
-                onClick = onButtonClick,
-                modifier = Modifier.background(MaterialTheme.colors.secondary, CircleShape)
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = buttonDescription,
-                    tint = MaterialTheme.colors.primary
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun PrettyListContent(content: List<@Composable () -> Unit>) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(8.dp),
-        modifier = Modifier.verticalScroll(rememberScrollState())
-    ) {
-        for (element in content) {
-            Box(
-                modifier = Modifier
-                    .background(MaterialTheme.colors.secondaryVariant, RoundedCornerShape(8.dp))
-                    .padding(8.dp)
-                    .fillMaxWidth()
-            ) {
-                element()
-            }
-        }
-    }
 }
 
 @Composable
@@ -314,23 +206,6 @@ fun FriendCard(friend: Account, onRemoveFriend: (Account) -> Unit) {
                 tint = MaterialTheme.colors.primary
             )
         }
-    }
-}
-
-@Composable
-fun PrettyList(name: String,
-               showButton: Boolean,
-               buttonDescription: String = "",
-               onButtonClick: () -> Unit = {},
-               content: List<@Composable () -> Unit>) {
-    Column(
-        modifier = Modifier
-            .background(MaterialTheme.colors.background, RoundedCornerShape(16.dp))
-            .padding(8.dp)
-            .fillMaxWidth()
-    ) {
-        ListHeader(name, showButton, buttonDescription, onButtonClick)
-        PrettyListContent(content)
     }
 }
